+++ ReportCard/Core/ExploredStatsController.cs
@@ -43,9 +43,9 @@ public static class ExploredStatsController {
     if (IsStatsPanelValid()) {
       StatsPanel.Panel.SetActive(true);
       StatsPanel.ResetStatsList();
+      StatsPanel.UpdateStatus(string.Empty);
 
-      TextMeshProUGUI label = StatsPanel.AddStatLabel();
-      _exploredStats.Generate(Minimap.m_instance, label, UpdateStatsPanel);
+      _exploredStats.Generate(Minimap.m_instance, StatsPanel.StatusLabel, UpdateStatsPanel);
     }
   }
 
@@ -59,10 +59,33 @@ public static class ExploredStatsController {
     if (IsStatsPanelValid()) {
       StatsPanel.ResetStatsList();
       StatsPanel.UpdateStatsList(explorerStats);
+      StatsPanel.UpdateStatus(GenerateStatusString(Player.m_localPlayer, ZNet.m_world, EnvMan.s_instance));
     }
   }
 
+  static string GenerateStatusString(Player player, World world, EnvMan envManager) {
+    if (!player || world == default || !envManager) {
+      return string.Empty;
+    }
+
+    return
+        $"<align=left>{player.GetPlayerName()}<line-height=0>\n"
+            + $"<align=right>{world.m_name}<line-height=1em>\n"
+            + $"<align=left><color=#D3D3D3>{player.GetPlayerID()}</color><line-height=0>\n"
+            + $"<align=right>Day <color=#FFD600>{envManager.GetCurrentDay()}</color><line-height=1em>";
+  }
+
   public static LabelButton StatsButton { get; private set; }
+
+  public static bool IsStatsButtonValid() {
+    return StatsButton?.Container;
+  }
+
+  public static void ToggleStatsButton(bool toggleOn) {
+    if (IsStatsButtonValid()) {
+      StatsButton.Container.SetActive(toggleOn);
+    }
+  }
 
   public static void CreateStatsButton(Minimap minimap) {
     StatsButton = new(minimap.m_largeRoot.transform);