+++ ReportCard/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 1.3.0
+
+  * Added new config-option `showStatsButton` under `ExploredStatsPanel`.
+  * `ExploredStatsPanel` now shows at the bottom: player name, player ID, world name, world day.
+
 ### 1.2.0
 
   * Added new config option `showRawValues` for the `ExploredStatsPanel` to toggle showing raw values on that panel.